move `CrossPlatformRequest` to types package
